requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.2
djangorestframework==3.15.1
sqlparse==0.4.4
tzdata==2024.1
